book_management/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.4
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.7.0
drf-yasg==1.21.8
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pymongo==4.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
